refactor project DB
