+++ services/web/server/src/simcore_service_webserver/projects/projects_api.py
@@ -120,7 +120,9 @@ from ..users.preferences_api import (
 from ..wallets import api as wallets_api
 from ..wallets.errors import WalletNotEnoughCreditsError
 from ..workspaces import _workspaces_db as workspaces_db
-from . import _crud_api_delete, _nodes_api
+from . import _crud_api_delete
+from . import _db_v2 as project_db_v2
+from . import _nodes_api
 from ._access_rights_api import (
     check_user_project_permission,
     has_user_project_access_rights,
@@ -289,7 +291,8 @@ async def patch_project(
             raise ProjectOwnerNotFoundInTheProjectAccessRightsError
 
     # 4. Patch the project
-    await db.patch_project(
+    await project_db_v2.patch_project(
+        app=app,
         project_uuid=project_uuid,
         new_partial_project_data=_project_patch_exclude_unset,
     )
