+++ services/web/server/tests/unit/with_dbs/03/test_project_db.py
@@ -201,7 +201,7 @@ async def insert_project_in_db(
             for group_id, permissions in _access_rights.items():
                 await update_or_insert_project_group(
                     client.app,
-                    new_project["uuid"],
+                    project_id=new_project["uuid"],
                     group_id=int(group_id),
                     read=permissions["read"],
                     write=permissions["write"],
